fixed errors

+++ app/src/main/java/com/adrian/airquality/data/remote/AirQualityDataDto.kt
@@ -5,21 +5,21 @@ import com.squareup.moshi.Json
 data class AirQualityDataDto(
     val time : List<String>,
     @field:Json(name = "pm10")
-    val pm10s : List<Int>,
+    val pm10s : List<Double>,
     @field:Json(name = "pm2_5")
-    val pm2_5s : List<Int>,
+    val pm2_5s : List<Double>,
     @field:Json(name = "carbon_monoxide")
-    val carbon_monoxides : List<Int>,
+    val carbon_monoxides : List<Double>,
     @field:Json(name = "nitrogen_dioxide")
-    val nitrogen_dioxides : List<Int>,
+    val nitrogen_dioxides : List<Double>,
     @field:Json(name = "sulphur_dioxide")
-    val sulphur_dioxides : List<Int>,
+    val sulphur_dioxides : List<Double>,
     @field:Json(name = "ozone")
-    val ozones : List<Int>,
+    val ozones : List<Double>,
     @field:Json(name = "aerosol_optical_depth")
     val aerosol_optical_depths : List<Double>,
     @field:Json(name = "dust")
-    val dusts : List<Int>,
+    val dusts : List<Double>,
     @field:Json(name = "uv_index")
     val uv_indexes : List<Double>,
     @field:Json(name = "uv_index_clear_sky")
