+++ app/src/main/java/com/adrian/airquality/domain/AirQualityData.kt
@@ -4,14 +4,14 @@ import java.time.LocalDateTime
 
 data class AirQualityData(
     val time: LocalDateTime,
-    val pm10: Int,
-    val pm25: Int,
-    val carbonMonoxide: Int,
-    val nitrogenDioxide: Int,
-    val sulphurDioxide: Int,
-    val ozone: Int,
+    val pm10: Double,              // Updated to Double
+    val pm25: Double,              // Updated to Double
+    val carbonMonoxide: Double,    // Updated to Double
+    val nitrogenDioxide: Double,   // Updated to Double
+    val sulphurDioxide: Double,    // Updated to Double
+    val ozone: Double,             // Updated to Double
     val aerosolOpticalDepth: Double,
-    val dust: Int,
+    val dust: Double,              // Updated to Double
     val uvIndex: Double,
     val uvIndexClearSky: Double
 )
